requirements.txt
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# GitHub scraper dependencies
requests==2.31.0

# LinkedIn scraper dependencies
selenium==4.15.2
webdriver-manager==4.0.1

# Additional dependencies
python-multipart==0.0.6

# For Render deployment compatibility
gunicorn==21.2.0 